telegram_bot/requirements.txt
python-telegram-bot
pydub
SpeechRecognition
apscheduler
transformers
deep-translator
nest-asyncio


